Turn KamiGuiScreen into a `class`
This cleans up the `parent` state, no longer needing to dirtily set it before opening the gui screen. Also removed the `invoke` overload, opting for a properly named method instead, which is less unsightly.

Also added one line to ClickGui's bind listener which cancels the key event, to mark it as handled. This also prevents it from propagating through to the currently shown screen (and thus, potentially causing a second action).