Don't open the ClickGui anywhere if an element expecting input is focused
This introduces a few `expectingInput` extension properties that test various elements (Screens, recipe books and widgets) for whether or not they're expecting input. These checks are limited to the ones in vanilla and do not attempt to cover any third party widgets that might expect input, as minecraft does not have an API to communicate that information.
Precisely, the covered elements are sign editing screens, book editing screens, creative inventory screens, furnace screens with an open recipe book, crafting screen with an open recipe book, or any other screen in which a `TextFieldWidget` is active (this may include third party screens!)

Closes #505