ESP: resolve issue with rendering backedModels that do not have a direction.
Some BakedModels only render when the direction is null. This is the case for Nether Portals, Fire, etc...

This is done in the same way that render{Smooth,Flat} in BlockModelRenderer excluding the light calculations.

End Portals use a different rendering technique that still needs to be resolved to get it to work.