fix: Invalidate cached widgets
Fiber (reasonably) assumes that the data structures you give it are immutable - which it also enforces (using immutable collections) for the default list ConfigTypes, etc.

However, in the case of enabled widgets, this immutability is a bit of a roadblock. Thus, we make them mutable, but that means fiber is not guaranteed to reserialize the values (as, to fiber, it already did - it's the same object after all.)

The fix: just disable caching on these settings. They are always reserialized.